after_2/requirements.txt
Flask
python-dotenv
pydantic